Remove deprecated functions and delete TimeSeriesUtil
