+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/dashboard/resources/AnomalyFunctionResource.java
@@ -1,8 +1,7 @@
 package com.linkedin.thirdeye.dashboard.resources;
 
-import com.linkedin.pinot.pql.parsers.utils.Pair;
-import com.linkedin.thirdeye.anomaly.detection.TimeSeriesUtil;
-import com.linkedin.thirdeye.api.DimensionKey;
+import com.linkedin.thirdeye.anomaly.detection.AnomalyDetectionInputContext;
+import com.linkedin.thirdeye.anomaly.detection.AnomalyDetectionInputContextBuilder;
 import com.linkedin.thirdeye.api.DimensionMap;
 import com.linkedin.thirdeye.api.MetricTimeSeries;
 import com.linkedin.thirdeye.client.DAORegistry;
@@ -116,19 +115,24 @@ public class AnomalyFunctionResource {
     // TODO: replace this with Job/Task framework and job tracker page
     BaseAnomalyFunction anomalyFunction = anomalyFunctionFactory.fromSpec(anomalyFunctionSpec);
 
-    List<Pair<Long, Long>> startEndTimeRanges = anomalyFunction.getDataRangeIntervals(startTime, endTime);
+    AnomalyDetectionInputContextBuilder anomalyDetectionInputContextBuilder =
+        new AnomalyDetectionInputContextBuilder(anomalyFunctionFactory);
 
-    Map<DimensionKey, MetricTimeSeries> dimensionKeyMetricTimeSeriesMap =
-        TimeSeriesUtil.getTimeSeriesForAnomalyDetection(anomalyFunctionSpec, startEndTimeRanges);
+    DateTime windowStart = new DateTime(startTime);
+    DateTime windowEnd = new DateTime(endTime);
+    anomalyDetectionInputContextBuilder.init(anomalyFunctionSpec)
+        .fetchTimeSeriesData(windowStart, windowEnd)
+        .fetchSaclingFactors(windowStart, windowEnd);
+    AnomalyDetectionInputContext anomalyDetectionInputContext = anomalyDetectionInputContextBuilder.build();
+
+    Map<DimensionMap, MetricTimeSeries> dimensionKeyMetricTimeSeriesMap =
+        anomalyDetectionInputContext.getDimensionKeyMetricTimeSeriesMap();
 
     List<RawAnomalyResultDTO> anomalyResults = new ArrayList<>();
     List<RawAnomalyResultDTO> results = new ArrayList<>();
-    List<String> collectionDimensions = DAO_REGISTRY.getDatasetConfigDAO()
-        .findByDataset(anomalyFunctionSpec.getCollection()).getDimensions();
 
-    for (Map.Entry<DimensionKey, MetricTimeSeries> entry : dimensionKeyMetricTimeSeriesMap.entrySet()) {
-      DimensionKey dimensionKey = entry.getKey();
-      DimensionMap dimensionMap = DimensionMap.fromDimensionKey(dimensionKey, collectionDimensions);
+    for (Map.Entry<DimensionMap, MetricTimeSeries> entry : dimensionKeyMetricTimeSeriesMap.entrySet()) {
+      DimensionMap dimensionMap = entry.getKey();
       if (entry.getValue().getTimeWindowSet().size() < 2) {
         LOG.warn("Insufficient data for {} to run anomaly detection function", dimensionMap);
         continue;
